Reference

Parameters

ParameterLinearValue

DifferentiableMetabolism.ParameterLinearValue — Type

struct ParameterLinearValue <: ConstraintTrees.Value

An extension of ConstraintTrees.LinearValue where the weights are parameters.

ParameterLinearValues can be combined additively and multiplied by real-number constants.

Multiplying two ParameterLinearValues yields a quadratic form (in a ParameterQuadraticValue).

Fields

idxs::Vector{Int64}

weights::Vector{Symbolics.Num}

ParameterQuadraticValue

DifferentiableMetabolism.ParameterQuadraticValue — Type

struct ParameterQuadraticValue <: ConstraintTrees.Value

An extension of ConstraintTrees.QuadraticValue where the weights are parameters.

Behaves similarly to ConstraintTrees.QuadraticValue. Thus, the cleanest way to construct a ParameterQuadraticValue is to multiply two ParameterLinearValues.

Fields

idxs::Vector{Tuple{Int64, Int64}}

weights::Vector{Symbolics.Num}

ParameterBound

DifferentiableMetabolism.ParameterBetween — Type

mutable struct ParameterBetween <: ConstraintTrees.Bound

Representation of an "interval" bound where the lower and upper bound values are parameters. Since Symbolics.Num is a subtype of Real, the bounds could also be any real number, but they are converted by the constructors to Symbolics.Nums.

Fields

lower::Symbolics.Num

upper::Symbolics.Num

DifferentiableMetabolism.ParameterEqualTo — Type

mutable struct ParameterEqualTo <: ConstraintTrees.Bound

Representation of an "equality" bound, where the bound value is a parameter. Since Symbolics.Num is a subtype of Real, the bound could also be any real number, but it is converted by the constructor to a Symbolics.Num.

Fields

equal_to::Symbolics.Num

ParameterIsozyme

DifferentiableMetabolism.ParameterIsozyme — Type

mutable struct ParameterIsozyme

A parameterized isozyme struct which includes parameters in the kcat_forward, and the kcat_backward. If the reaction does not have a turnover number,nothing can be used.

Fields

gene_product_stoichiometry::Dict{String, Float64}

kcat_forward::Union{Nothing, Symbolics.Num}

kcat_reverse::Union{Nothing, Symbolics.Num}

Parameterized models

Kinetic models

DifferentiableMetabolism.build_kinetic_model — Method

build_kinetic_model(
    model::AbstractFBCModels.AbstractFBCModel;
    reaction_isozymes,
    gene_product_molar_masses,
    capacity
)

Essentially an enzyme constrained metabolic model, but the kcat can be a arbitrary symbolic function.

Solving models

DifferentiableMetabolism.constraint_matrix_vector — Method

constraint_matrix_vector(
    eqs,
    m,
    parameters
) -> Tuple{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseVector{Float64, Int64}}

Builds a matrix representation of bounds.

DifferentiableMetabolism.optimization_model_with_parameters — Method

optimization_model_with_parameters(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    parameters::Dict{Symbolics.Num, Float64};
    objective,
    optimizer,
    sense
)

Construct a JuMP model by substituting parameters into the model, m. Set the objective and the optimizer, as well as the sense similar to COBREXA.optimization_model.

Converts all inequality constraints to the form A * x ≤ b.

DifferentiableMetabolism.optimized_constraints_with_parameters — Method

optimized_constraints_with_parameters(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    parameters::Dict{Symbolics.Num, Float64};
    modifications,
    objective,
    optimizer,
    sense
)

Solve a model, m, by forwarding arguments to optimization_model_with_parameters.

Optionally, set optimizer attributes with modifications. If the model does not solve successfully return nothing. Otherwise, return a tuple of the solution tree, and vectors containing the values of the primal variables, the equality constraint dual variables.

These duals are ordered according to the constraint output of calling equality_constraints and inequality_constraints respectively.

Pruning models

DifferentiableMetabolism.prune_gene_product_molar_masses — Method

prune_gene_product_molar_masses(
    gene_product_molar_masses,
    ec_solution,
    gene_zero_tol
) -> Dict

TODO

DifferentiableMetabolism.prune_model — Method

prune_model(
    model,
    ec_solution,
    flux_zero_tol,
    gene_zero_tol
) -> AbstractFBCModels.CanonicalModel.Model

Prune away reactions, metabolites, and genes from a model using ec_solution, which is the result of an enzyme constrained kinetic simulation. Fluxes and gene product concentrations smaller than flux_zero_tol, gene_zero_tol are removed. Metabolites that do not take part in the remaining reactions are also removed.

DifferentiableMetabolism.prune_reaction_isozymes — Method

prune_reaction_isozymes(
    reaction_isozymes,
    ec_solution,
    gene_zero_tol
) -> Dict

TODO

Differentiation

DifferentiableMetabolism.differentiate — Method

differentiate(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    objective::ConstraintTrees.Value,
    x_vals::Vector{Float64},
    eq_dual_vals::Vector{Float64},
    ineq_dual_vals::Vector{Float64},
    parameter_values::Dict{Symbolics.Num, Float64},
    parameters::Vector{Symbolics.Num};
    scale
) -> Tuple{Any, Any}

Differentiate a model m with respect to parameters which take on values parameter_values in the optimal solution with respect to the objective. The primal variables x_vals, and the dual variable values eq_dual_vals and ineq_dual_vals need to be supplied.

Internally, primal variables with value abs(x) ≤ primal_zero_tol are removed from the computation, and their sensitivities are not calculated.

DifferentiableMetabolism.findall_indeps_qr — Method

findall_indeps_qr(A) -> Any

TODO

DifferentiableMetabolism.variable_order — Method

variable_order(m) -> Any

TODO

Internals

Constraint tree extensions

Optimization problem builders

DifferentiableMetabolism.constrained — Method

constrained(x) -> Bool

TODO

DifferentiableMetabolism.equality_constraints — Method

equality_constraints(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Vector{Tuple{Union{ConstraintTrees.LinearValue, ParameterLinearValue}, Symbolics.Num}}

Return all the equality constraints of m as a tuple ({Parameter}LinearValue, value) representing {P}LV == value for each entry.

DifferentiableMetabolism.get_equality_constraints — Method

get_equality_constraints(
    c::ConstraintTrees.Constraint,
    sink
) -> Any

TODO

DifferentiableMetabolism.get_equality_constraints — Method

get_equality_constraints(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    sink
) -> Any

TODO

DifferentiableMetabolism.get_inequality_constraints — Method

get_inequality_constraints(
    c::ConstraintTrees.Constraint,
    sink
) -> Any

TODO

DifferentiableMetabolism.get_inequality_constraints — Method

get_inequality_constraints(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    sink
) -> Any

TODO

DifferentiableMetabolism.inequality_constraints — Method

inequality_constraints(
    m::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Vector

Return all the inequality constraints of m as a tuple of bounds converted to the form ({Parameter}LinearValue, upper) representing {P}LV ≤ upper for each entry.

Symbolics extensions