Reference

Solver implementation

DifferentiableMetabolism.constraint_matrix_vector — Method

constraint_matrix_vector(
    eqs,
    m,
    parameters
) -> Tuple{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseVector{Float64, Int64}}

Builds a matrix representation of bounds.

DifferentiableMetabolism.optimization_model — Method

optimization_model(
    m::ConstraintTrees.ConstraintTree,
    parameters::Dict{Symbol, Float64};
    objective,
    optimizer,
    sense
)

Construct a JuMP model by substituting parameters into the model, m. Set the objective and the optimizer, as well as the sense similar to COBREXA.optimization_model.

Converts all inequality constraints to the form A * x ≤ b.

DifferentiableMetabolism.optimized_values — Method

optimized_values(
    model::ConstraintTrees.ConstraintTree,
    parameters::Dict{Symbol, Float64};
    optimizer,
    settings,
    objective,
    sense
)

Solve a model using optimizer by substituting in parameters. Optional arguments are the same as in COBREXA.

If the model does not solve successfully return nothing. Otherwise, return a named tuple of the solution tree, and vectors containing the values of the primal variables, the dual variables.

These duals are ordered according to the constraint output of calling equality_constraints and inequality_constraints respectively.

DifferentiableMetabolism.equality_constraints — Method

equality_constraints(
    m::ConstraintTrees.ConstraintTree
) -> Vector{Tuple{Union{ConstraintTrees.LinearValueT{FastDifferentiation.Node}, ConstraintTrees.LinearValue}, FastDifferentiation.Node}}

Return all the equality constraints of m as a tuple (constraint, bound) representing constraint == bound for each entry.

DifferentiableMetabolism.get_equality_constraints — Method

get_equality_constraints(
    c::ConstraintTrees.Constraint,
    sink
) -> Any

Helper function for equality_constraints.

DifferentiableMetabolism.get_equality_constraints — Method

get_equality_constraints(
    m::ConstraintTrees.ConstraintTree,
    sink
) -> Any

Helper function for equality_constraints.

DifferentiableMetabolism.get_inequality_constraints — Method

get_inequality_constraints(
    c::ConstraintTrees.Constraint,
    sink
) -> Any

Helper function for inequality_constraints.

DifferentiableMetabolism.get_inequality_constraints — Method

get_inequality_constraints(
    m::ConstraintTrees.ConstraintTree,
    sink
) -> Any

Helper function for inequality_constraints.

DifferentiableMetabolism.inequality_constraints — Method

inequality_constraints(
    m::ConstraintTrees.ConstraintTree
) -> Vector

Return all the inequality constraints of m as a tuple of bounds converted to the form (constraint, upper) representing constraint ≤ upper for each entry.

DifferentiableMetabolism.is_constrained — Method

is_constrained(x) -> Bool

Check if the argument is constrained (i.e. bound not infinity) or not.

DifferentiableMetabolism.substitute — Method

substitute(x::FastDifferentiation.Node, lookup) -> Any

Straightforward recursive evaluator for Exs.

Overloads

DifferentiableMetabolism.variable_order — Method

variable_order(m) -> Any

Return the names of variables in m.

NOTE: this function assumes that all variables are bounded explicitly in the model. nothing bounds are ignored.

Pruning models

DifferentiableMetabolism.prune_model — Method

prune_model(
    model::AbstractFBCModels.CanonicalModel.Model,
    solution_fluxes,
    solution_gene_product_amounts,
    reaction_isozymes,
    solution_isozyme_forward_amounts,
    solution_isozyme_reverse_amounts,
    flux_zero_tol::Float64,
    gene_zero_tol::Float64
) -> Tuple{AbstractFBCModels.CanonicalModel.Model, Dict}

Prune away reactions, metabolites, and genes from a model using solution. Additionally, adjust reaction_isozymes using prune_reaction_isozymes to account for the changed directions of the fluxes.

Fluxes and gene product concentrations smaller than flux_zero_tol, gene_zero_tol are removed. Metabolites that do not take part in the remaining reactions are also removed.

Note, in COBREXA reverse variables are always derived quantities, hence making everything forward orientated ensures that the variables of the whole model are just these.

DifferentiableMetabolism.prune_reaction_isozymes — Method

prune_reaction_isozymes(
    reaction_isozymes,
    solution_isozyme_forward_amounts,
    solution_isozyme_reverse_amounts,
    solution_fluxes,
    gene_zero_tol::Float64
) -> Dict

Return a freshly allocated dictionary mapping reaction IDs to isozyme structs. Note, all reactions running in reverse have their kcats swapped, since prune_model changes all fluxes to be positive.

Example

prune_reaction_isozymes(reaction_isozymes, solution.isozyme_forward_amounts, solution.isozyme_reverse_amounts, solution.fluxes, 1e-6)

Differentiation of pruned models

DifferentiableMetabolism.differentiate_prepare_kkt — Method

differentiate_prepare_kkt(
    m::ConstraintTrees.ConstraintTree,
    objective::ConstraintTrees.Value,
    parameters::Vector{Symbol}
) -> Tuple{Tuple{SparseArrays.SparseMatrixCSC{FastDifferentiation.Node, Int64}, SparseArrays.SparseMatrixCSC{FastDifferentiation.Node, Int64}, Vector{FastDifferentiation.Node}, Vector{FastDifferentiation.Node}, Vector{FastDifferentiation.Node}, Vector{Symbol}}, Any}

Prepare a model m with objective for differentiation with respect to parameters.

This is the most time consuming aspect of differentiation. It pays off to do this separately if the same model will be differentiated multiple times.

DifferentiableMetabolism.differentiate_solution — Method

differentiate_solution(
    ,
    primal_vals::Vector{Float64},
    eq_dual_vals::Vector{Float64},
    ineq_dual_vals::Vector{Float64},
    parameter_values::Dict{Symbol, Float64};
    scale
) -> Any

Differentiate a solution of a model. The first argument is the output of differentiate_prepare_kkt, and is a tuple of the deconstructed model. The following arguments (primal_vals, eq_dual_vals, ineq_dual_vals) are outputs of optimized_values. parameter_values

DifferentiableMetabolism.findall_indeps_qr — Method

findall_indeps_qr(A) -> Any

Return all linearly dependent constraints in A, using the QR decomposition.